+++ screens/lista_cliente.html
@@ -71,6 +71,6 @@
       </div>
     </main>
 
-    <script src="../service/client-service.js"></script>
+    <script type="module" src="../controllers/client_controllers.js"></script>
   </body>
 </html>
